% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envir_functions.R
\name{calc_overwaterdist_point}
\alias{calc_overwaterdist_point}
\title{Function to calculate overwater distance from a single location or set of locations}
\usage{
calc_overwaterdist_point(
  landmask,
  coords,
  out_dir,
  dist_unit = c("m", "km"),
  plot = TRUE,
  overwrite = FALSE
)
}
\arguments{
\item{landmask}{A SpatRaster where land is NA and water == 1.}

\item{coords}{Data frame with columns labelled 'location', 'x', and 'y'. Where location is the
location name used for saving each distance raster, x and y are the location coordinates in the same units as landmask.}

\item{out_dir}{File path to the directory where slope raster will be saved. String.}

\item{dist_unit}{Unit for the distances One of "m" or "km".}

\item{plot}{Should the distance rasters be plotted. Logical.}

\item{overwrite}{Should existing files be overwritten. Logical.}
}
\value{
A separate ncdf file with overwater distance calculate for each record in coords. Output saved to out_dir.
}
\description{
Function to calculate overwater distance from a single location or set of locations
}
\examples{

# Get bathymetry
get_bathymetry(out_dir = 'tmp',
               region = c(-68, -63, 44, 46),
               overwrite = FALSE,
               stride = 1,
               vars = NULL)

r <- terra::rast('tmp/bathymetry.nc')
lm <- terra::ifel(r >=0, NA, 1)

my_cols <- data.frame(location = c('loc1', 'loc2'),
                      x = c(-66.718, -66.387),
                      y = c(44.702, 44.240)
)

calc_overwaterdist_point(landmask = lm,
                         coords = my_cols,
                         out_dir = 'tmp/colony_dist',
                         dist_unit = c('km'),
                         plot = TRUE,
                         overwrite = TRUE)

unlink('tmp', recursive = TRUE)

}
