% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envir_functions.R
\name{check_colony}
\alias{check_colony}
\title{Function to check if colony coordinates touch water and, if not, adjust coordinates
to nearest water point. Internal function for within get_colonydist}
\usage{
check_colony(landmask, coords)
}
\arguments{
\item{landmask}{A SpatRaster object. Land is converted to NA and water to 1 within function.}

\item{coords}{Data frame with columns labelled 'location', 'x', and 'y'. Where location is the
location name used for saving each distance raster, x and y are the location coordinates in the same units as landmask.}
}
\value{
An adjusted colony coordinate data frame to be used in get_colonydist.
}
\description{
Function to check if colony coordinates touch water and, if not, adjust coordinates
to nearest water point. Internal function for within get_colonydist
}
\examples{

# Get bathymetry
get_bathymetry(out_dir = 'tmp',
               region = c(-68, -63, 44, 46),
               overwrite = FALSE,
               stride = 1,
               vars = NULL)

landmask <- terra::rast('tmp/bathymetry.nc')

my_cols <- data.frame(location = c('loc1', 'loc2'),
                      x = c(-66.718, -66.387),
                      y = c(44.702, 44.240)
)

check_colony(landmask = landmask,
             coords = my_cols)

}
