% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cms_functions.R
\docType{data}
\name{cms_table}
\alias{cms_table}
\title{Copernicus Marine Service product table}
\format{
A data frame
\describe{
\item{variable}{CMS name of the variable that will be downloaded}
\item{frequency}{Temporal frequency of the data to be downloaded}
\item{product_id}{CMS product ID}
\item{dataset_id}{CMS dataset ID}
\item{dataset_name}{Name used when saving outputs from this dataset}
\item{min_date}{Earliest date available in this dataset}
\item{max_date}{Latest date available in this dataset}
}
}
\usage{
cms_table
}
\description{
Metadata information about Copernicus Marine Service datasets that can be downloaded using get_cms_data()
}
\author{
Allison Patterson \email{allison.patterson@ec.gc.ca}
}
\keyword{datasets}
