% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotspot_functions.R
\name{erase_poly}
\alias{erase_poly}
\title{Erase parts of polygons}
\usage{
erase_poly(x, y)
}
\arguments{
\item{x}{An sf POLYGONS object}

\item{y}{An sf POLYGONS object}
}
\value{
An sf POLYGONS object
}
\description{
A helper function that errases any part of x that overlaps with y
}
\examples{

library(ggplot2)
d <- coei
d <- sf::st_as_sf(d, coords = c('longitude', 'latitude'), crs = 4326)
d <- sf::st_transform(d,
                      crs = '+proj=laea +lat_0=50 +lon_0=-65 +x_0=0 +y_0=0 +datum=WGS84 +units=m +no_defs')

# get coast layer
coast <- rnaturalearth::ne_countries(scale = 10, return = 'sf')
coast <- sf::st_transform(coast, crs = sf::st_crs(d))
coast_crop <- sf::st_crop(coast, sf::st_buffer(d, dist = 10000))

bbox_poly <-  d |>
sf::st_buffer(dist = 10000) |>
  sf::st_bbox() |>
  sf::st_as_sfc()

p <- erase_poly(bbox_poly, coast_crop)
}
