% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cms_functions.R
\name{get_cms_data}
\alias{get_cms_data}
\title{Download subsetted data from Copernicus Marine Service (CMS)}
\usage{
get_cms_data(
  my_env,
  out_dir,
  var = "thetao",
  prod = "cmems_mod_glo_phy_hindcast",
  freq = c("monthly", "daily", "static"),
  region = c(-68, -63, 44, 46),
  date_min = Sys.Date() - 1,
  date_max = Sys.Date(),
  depth_min = 0,
  depth_max = 0,
  overwrite = FALSE
)
}
\arguments{
\item{my_env}{Path to a virtual python environment containing the Copernicus Marine Toolbox}

\item{out_dir}{File path to the directory where raster will be saved. String.}

\item{var}{Variable name to be downloaded from CMS}

\item{prod}{CMS product ID}

\item{freq}{Frequency of the dataset to download. Currently supports: daily, monthly, and static.}

\item{region}{Bounding box for region as: xmin, xmax, ymin, ymax.}

\item{date_min}{The earliest date to include, formatted as a date class.}

\item{date_max}{The latest date to include, formatted as a date class.}

\item{depth_min}{For datasets with a depth component, the minimum depth value to include. Default is 0.}

\item{depth_max}{For datasets with a depth component, the maximum depth value to include. Default is 0.}

\item{overwrite}{TRUE/FALSE. Should existing files be overwritten? Default is FALSE.}
}
\value{
Netcdf file(s) saved to out_dir.
}
\description{
Download subsetted data from Copernicus Marine Service (CMS)
}
\details{
This is a wrapper function to use the subset function of the Copernicus
Marine Toolbox directly from R with reticulate. CMS provides access to many useful
oceanographic datasets. This function can be used to download certain commonly
used datasets programmatically. Because the full data files cover the entire globe,
often at multiple depth layers making it impractical to download the full original files
for large time ranges. The subsetting tool allows us to request custom spatial, temporal, and depth
extents.

This function relies on a data table that is pre-populated with required fields for the
datasets that are most commonly used in species distribution modelling: Global
Ocean Physics and Global Ocean Biogeochemistry. Other datasets could be added on request.
The hindcasast nd forecast models have different dataset IDs and cover different date ranges,
so this function requests data from the correct datasets for the temporal extent requested.
If the request covers multiple datasets, then a netcdf will be saved for each dataset.

Before running this function for the first time you will need to create a Python
virtual environment with the copernicus marine toolbox installed. This can be
done using the get_cms_env() function. This environment could be saved outside
of an R project and called by any project that requires CMS data.
}
\examples{

data('cms_table')
cms_table

# not run
if (FALSE) {
get_cms_data(
   my_env = cm_env,
   out_dir = 'tmp',
   var = 'so',
   freq = 'monthly',
   region = c(-68, -63, 44, 46),
   date_min = Sys.Date() + -31,
   date_max = Sys.Date() + 7,
   depth_min = 0.49402499198913574,
   depth_max = 0.49402499198913574
   )
}
}
\seealso{
make_cms_envir()
}
