% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envir_functions.R
\name{get_coastdist}
\alias{get_coastdist}
\title{Function to calculate slope from a bathymetry file}
\usage{
get_coastdist(
  bath_file,
  out_dir,
  mask_land = 0,
  dist_unit = c("m", "km"),
  overwrite = FALSE
)
}
\arguments{
\item{bath_file}{File path to saved bathymetry raster. String.}

\item{out_dir}{File path to the directory where slope raster will be saved. String.}

\item{mask_land}{Value of elevation that will be used to mask out land. Integer. Default is 0}

\item{dist_unit}{Unit for the distances One of "m" or "km".}

\item{overwrite}{Should existing files be overwritten. Logical.}
}
\value{
A netcdf file named 'coastdist.nc' saved to out_dir
}
\description{
Function to calculate slope from a bathymetry file
}
\details{
This function is a wrapper for terra::costDist(). This function requires a bathymetry
file as input, where elevation values considered to be on land have been converted
to NA.

If the input raster includes on land elevations, then the mask_land argument,
can be used to mask values greater than mask_land to NA.
}
\examples{
get_bathymetry(out_dir = 'tmp',
              region = c(-68, 44, -63, 46),
              overwrite = FALSE,
              stride = 1,
              mask_land = 0)
get_coastdist(bath_file = 'tmp/bathymetry.nc',
             out_dir = 'tmp',
             mask_land = 0,
             dist_unit = 'km',
             overwrite = TRUE)
r <- terra::rast('tmp/coastdist.nc')
terra::plot(r)
}
\seealso{
\code{\link[=get_bathymetry]{get_bathymetry()}}, \code{\link[terra:costDist]{terra::costDist()}}
}
