% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotspot_functions.R
\name{get_flocks}
\alias{get_flocks}
\title{Create flocks from census observations}
\usage{
get_flocks(dat, flock_spacing, search_dist = 100, np = 100, coast)
}
\arguments{
\item{dat}{An sf points object of survey observations. Must include a column named 'total' that records the number of birds seen per observation. Must be in a projected coordinate system.}

\item{flock_spacing}{Standard deviation in m for generating birds around previous observation}

\item{search_dist}{Standard deviation in m for generating birds when trying to move birds over land onto water}

\item{np}{Number of random points to generate when trying to move points on land onto water.}

\item{coast}{An sf polygon object of land, which will be used to ensure that birds are always lcoated in water. These polygons could be configured to exclude birds from any areas (e.g. exclude land birds from ocean).}
}
\value{
An sf POINT object with one row for each bird in observed flocks. If
some birds are still overlapping with the coast polygons a warning will be given.
The output will include all fields from dat, plus a flock_id field indicating
which row the observation is associated with.
}
\description{
Create flocks from census observations
}
\details{
In aerial census surveys of seaducks, a single point is recorded for each flock observed.
This function is the first step in a workflow for identifying hotspots from census surveys.
The function generates a single row for every individual in the flock. Because
very large flocks would have occupied more geographic space than small flocks, the
function adds individuals around the first observation using a random walk with the
flock_spacing argument determining the distribution of distances between subsequent points.

The coast argument allows users to specify areas (land) that are inaccessible to the
focal species. Any observations that end up overlapping the coast layer are moved to the closest
randomly generated point within search_distance of the overlapping point. Currently there is no
option to run this function without a 'coast' layer. The example uses the rnaturalearth::ne_countries()
layer, however in mnay applications for coastal species this shoreline may be too coarse and the
user should consider a layer wih higher resolution.

This function can be quite slow for large data sets. User should crop their coast layer to
focus on the area of interest (see example)
#'
}
\examples{

# format data
d <- coei
d <- sf::st_as_sf(d, coords = c('longitude', 'latitude'), crs = 4326)
d <- sf::st_transform(d,
                      crs = '+proj=laea +lat_0=50 +lon_0=-65 +x_0=0 +y_0=0 +datum=WGS84 +units=m +no_defs')

# get coast layer
coast <- rnaturalearth::ne_countries(scale = 10, return = 'sf')
coast <- sf::st_transform(coast, crs = sf::st_crs(d))
coast_crop <- sf::st_crop(coast, sf::st_buffer(d, dist = 10000))

# generate flocks from observation data
f <- get_flocks(dat = d,
                flock_spacing = 10,
                search_dist = 100,
                np = 100,
                coast = coast_crop)

sum(d$total) == nrow(f)
f

}
