% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envir_functions.R
\name{get_slope}
\alias{get_slope}
\title{Function to calculate slope from a bathymetry file}
\usage{
get_slope(
  bath_file,
  out_dir,
  neighbors = 8,
  slope_unit = c("degrees", "radians"),
  overwrite = FALSE
)
}
\arguments{
\item{bath_file}{File path to saved bathymetry raster. String.}

\item{out_dir}{File path to the directory where slope raster will be saved. String.}

\item{neighbors}{Number of neighbours to use in slope calculation. Either 4 or 8.}

\item{slope_unit}{Unit for the slope calculation. One of "degrees" or "radians".}

\item{overwrite}{Should existing files be overwritten. Logical.}
}
\value{
A netcdf file named 'slope.nc' saved to out_dir
}
\description{
Function to calculate slope from a bathymetry file
}
\details{
This function is a wrapper for terra::terrain(), which will calculate slope
and save a standardized file to ensure consistency with other outputs from PelagicMagic.
The function can be used directly with a bathymetry file as input or it can be galled with get_bathymetry()
when doing the initial download of bathymetry data.

See terra::terrain() for detailed description of how slope is calcualted.
}
\examples{
get_bathymetry(out_dir = 'tmp',
              region = c(-68, 44, -63, 46),
              overwrite = FALSE,
              stride = 1,
              mask_land = 0)
get_slope(bath_file = 'tmp/bathymetry.nc',
         out_dir = 'tmp',
         neighbors = 8,
         slope_unit = 'degrees',
         overwrite = FALSE)
r <- terra::rast('tmp/slope.nc')
terra::plot(r)
unlink('tmp', recursive = TRUE)

}
\seealso{
\code{\link[=get_bathymetry]{get_bathymetry()}}, \code{\link[terra:terrain]{terra::terrain()}}
}
