% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cms_functions.R
\name{make_cms_envir}
\alias{make_cms_envir}
\title{Create a Python virtual environment with copernicusmarine toolbox installed}
\usage{
make_cms_envir(my_env, version = "3.11.9", user = NULL, password = NULL)
}
\arguments{
\item{my_env}{The name of, or path to, a Python virtual environment. If this
name contains any slashes, the name will be interpreted as a path; if the name
does not contain slashes, it will be treated as a virtual environment within virtualenv_root().
When NULL, the virtual environment as specified by the RETICULATE_PYTHON_ENV
environment variable will be used instead. To refer to a virtual environment
in the current working directory, you can prefix the path with ⁠./\if{html}{\out{<name>}}⁠.}

\item{version}{Python version to use. String. Default is '3.11'.}

\item{user}{User name for copernicus marine service. String.}

\item{password}{Password for copernicus marine service. String.}
}
\value{
Creates a python virtual environment with the copernicusmarine toolbox installed on the path specified by my_env
}
\description{
Create a Python virtual environment with copernicusmarine toolbox installed
}
\details{
This function uses reticulate to create a Python virtual environment
and install the copernicusmarine toolbox. This environment can then be used to download
oceanographic data from Copernicus Marine Service.

If the function is run including the user name and password then login details
will be saved to the virtual environment on creation.
}
\examples{

# not run
if (FALSE) {
cm_env <- './tmp/cms_env'
make_cms_envir(my_env = cm_env)
reticulate::use_virtualenv(paste0(cm_env,'/'), required = TRUE)
cm <- reticulate::import("copernicusmarine")
}

}
