% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envir_functions.R
\name{pick_aea_projection}
\alias{pick_aea_projection}
\title{Pick a custom Albers Equal Area projection suitable for a given dataset}
\usage{
pick_aea_projection(
  dat,
  latcol = "lat",
  longcol = "lon",
  type = c("PROJ4", "WKT2")
)
}
\arguments{
\item{dat}{either an \pkg{sf} object or a dataframe with latitudes in
\code{latcol} and longitudes in \code{longcol}.}

\item{latcol, longcol}{names of \code{dat} columns containing latitudes
and longitudes, respectively. Ignored if \code{dat} is an \pkg{sf} object.}

\item{type}{type of formatting for the returned projection string.}
}
\value{
The projection description string in the chosen format.
}
\description{
Pick a custom Albers Equal Area projection suitable for a given dataset
}
\details{
Figures out what standard parallels to use for Albers Equal Area projection.
As per: \href{https://pro.arcgis.com/en/pro-app/latest/help/mapping/properties/albers.htm}{https://pro.arcgis.com/en/pro-app/latest/help/mapping/properties/albers.htm}
See also: \href{https://proj.org/en/6.2/operations/projections/aea.html}{https://proj.org/en/6.2/operations/projections/aea.html}

Parameters for Albers Equal Area projection:\cr\cr
lat_1: southern standard parallel - 1/6 of latitude range north of the southern limit\cr
lat_2: northern standard parallel - 1/6 of latitude range south of the northern limit\cr
lat_0: central parallel - midpoint of latitude range\cr
lon_0: central meridian - mean of longitude range (or maybe median?)\cr
}
\examples{

pick_aea_projection(quakes, latcol = "lat", longcol = "long")


library(sf)
quakes_sf <- st_as_sf(quakes, coords = c("long", "lat"), crs = 4326)
pick_aea_projection(quakes_sf)


}
