% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envir_functions.R
\name{get_bathymetry}
\alias{get_bathymetry}
\title{Function to download bathymetry data}
\usage{
get_bathymetry(
  out_dir = NULL,
  region,
  overwrite = FALSE,
  stride = 1,
  mask_land = NULL,
  add_slope = FALSE,
  neighbors = 8,
  slope_unit = c("degrees", "radians"),
  add_coastdist = FALSE,
  dist_unit = c("m", "km")
)
}
\arguments{
\item{out_dir}{File path to the directory where raster will be saved. String.}

\item{region}{Bounding box for region as xmin, ymin, xmax, ymax.}

\item{overwrite}{Should existing files be overwritten. Logical.}

\item{stride}{Rate at which to sample underlying raster resolution. Integer.}

\item{mask_land}{Value of elevation that will be used to mask out land. Integer. Default is NULL.}

\item{add_slope}{Should a layer for slope be calculated. Logical.}

\item{neighbors}{Number of neighbours to use in slope calculation. Either 4 or 8.}

\item{slope_unit}{Unit for the slope calculation. One of "degrees" or "radians".}

\item{add_coastdist}{Should a layer for distance to coastline be calculated. Logical.}

\item{dist_unit}{Unit for the distances One of "m" or "km".}
}
\value{
A netcdf file named 'bathymetry.nc' saved to out_dir. If add_slope or
add_coast are set to TRUE, then additional files named 'slope.nc' and 'coastdist.nc'
will be saved. Test git
}
\description{
Function to download bathymetry data
}
\details{
This function uses rerddap to download the GEBCO_2020 Grid, which is a
continuous terrain model for oceans and land at 15 arc-second intervals (0.004 deg).

Stride is passed to rerddap if you want to downsample the resolution of the
raster data before downloading. For example, a stride of 2 will only acquire
every second grid cell and a stride of 10 will acquire every 10th grid cell.

If mask_land is not null, all elevation values greater than mask_land will be converted to NA.
}
\examples{
# Just get bathymetry
get_bathymetry(out_dir = 'tmp',
              region = c(-68, 44, -63, 46),
              overwrite = FALSE,
              stride = 1,
              mask_land = 0,
              add_slope = FALSE,
              neighbors = 8,
              slope_unit = c('degrees','radians'),
              add_coastdist = FALSE,
              dist_unit = c('m','km'))
r <- terra::rast('tmp/bathymetry.nc')
terra::plot(r)

# Get bathymetry and calculate slope and distance from coast
get_bathymetry(out_dir = 'tmp',
              region = c(-68, 44, -63, 46),
              overwrite = FALSE,
              stride = 1,
              mask_land = 0,
              add_slope = TRUE,
              neighbors = 8,
              slope_unit = 'degrees',
              add_coastdist = TRUE,
              dist_unit = 'km')
r <- terra::rast('tmp/slope.nc')
terra::plot(r, main = terra::longnames(r))
r <- terra::rast('tmp/coastdist.nc')
terra::plot(r, main = terra::longnames(r))
unlink('tmp', recursive = TRUE)
}
\seealso{
\code{\link[=get_slope]{get_slope()}}, \code{\link[=get_coastdist]{get_coastdist()}}
}
