% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cms_functions.R
\name{make_cms_envir}
\alias{make_cms_envir}
\title{Create a Python vertual environment with copernicusmarine toolbox installed}
\usage{
make_cms_envir(my_env, force = FALSE, user = NULL, password = NULL)
}
\arguments{
\item{my_env}{The name of, or path to, a Python virtual environment. If this
name contains any slashes, the name will be interpreted as a path; if the name
does not contain slashes, it will be treated as a virtual environment within virtualenv_root().
When NULL, the virtual environment as specified by the RETICULATE_PYTHON_ENV
environment variable will be used instead. To refer to a virtual environment
in the current working directory, you can prefix the path with ⁠./\if{html}{\out{<name>}}⁠.}

\item{force}{Force recreating the environment specified by envname, even if
it already exists. If TRUE, the pre-existing environment is first deleted and
then recreated. Otherwise, if FALSE (the default), the path to the existing environment is returned.}

\item{user}{User name for copernicus marine service. String.}

\item{password}{Password for copernicus marine service. String.}
}
\value{
Creates a python virtual environment with the copernicusmarine toolbox installed on the path specified by my_env
}
\description{
Create a Python vertual environment with copernicusmarine toolbox installed
}
\details{
This function uses reticulate to create a Python virtual environment
and install the copernicusmarine toolbox. This toolbox can be used to download
oceanographic data from copernicus marine service.

If the function is run including the user name and password then login details
will be saved to the virtual environment on creation.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

cm_env <- 'tmp/cms_env'
make_cms_envir(my_env = cm_env, force = FALSE)
reticulate::use_virtualenv(paste0(my_env,'/'), required = TRUE)
cm <- reticulate::import("copernicusmarine")
cm$login()
\dontshow{\}) # examplesIf}
}
